adjusting delete\create function exceptions

+++ src/services/accounts-service/index.ts
@@ -1,4 +1,4 @@
-import { accountAlreadyExistsError } from './errors';
+import { accountAlreadyExistsError, invalidAccountIdError } from './errors';
 import accountsRepository from '@/repositories/accounts-repository';
 import { AccountEntity, Account } from '@/protocols/accounts';
 import accounts from '@/database/data';
@@ -11,10 +11,13 @@ async function listAccounts(): Promise<AccountEntity[]> {
 
 async function addAccount(bankAccount: Account): Promise<AccountEntity> {
   const existingAccount = await validateExistingAccount(bankAccount);
+  let idController = 0;
 
   if (existingAccount.length == 1) throw accountAlreadyExistsError();
 
-  const idController = accounts[accounts.length - 1].id;
+  if (accounts.length > 0) {
+    idController = accounts[accounts.length - 1].id;
+  }
 
   const newId = idController + 1;
 
@@ -29,6 +32,10 @@ async function addAccount(bankAccount: Account): Promise<AccountEntity> {
 }
 
 async function removeAccount(id: number): Promise<any> {
+  const existingId = await validateAccountId(id);
+
+  if (existingId.length == 0) throw invalidAccountIdError();
+
   const deletedAccount = await accountsRepository.deleteAccount(id);
 
   return deletedAccount;
@@ -44,6 +51,14 @@ async function validateExistingAccount(bankAccount: Account) {
   const filteredAccountNum = filteredAgency.filter((account) => account.accountNum === bankAccount.accountNum);
 
   return filteredAccountNum;
+}
+
+async function validateAccountId(id: number) {
+  const accountList = await accountsRepository.allAccounts();
+
+  const filteredAccountId = accountList.filter((account) => account.id === id);
+
+  return filteredAccountId;
 }
 
 export default {
